requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
Pillow
sqlalchemy
python-multipart
